aurus_reasoning/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64"]

[project]
name = "aurus_reasoning"
version = "0.1.0"
dependencies = [
  "nvidia-nat[langchain]~=1.4",
]
requires-python = ">=3.11,<3.14"
description = "Custom NeMo Agent Toolkit Workflow"
classifiers = ["Programming Language :: Python"]



[project.entry-points.'nat.components']
aurus_reasoning = "aurus_reasoning.register"